10 - Refactoring do Projeto - Aplicando o princípio na prática parte 3

+++ projsolid1app_carrinho_compras_b/src/Item.php
@@ -36,7 +36,7 @@ class Item
   /**
    * @return int
    */
-  public function getValor(): int
+  public function getValor(): float
   {
     return $this->valor;
   }
